move networkchainID state value from settings to wallet reducer
