+++ crosscash-extension/src/popup/store/settings/reducer.ts
@@ -1,10 +1,8 @@
 import { PayloadAction } from '@reduxjs/toolkit';
 
-import { MAINNET } from '../../../lib/constants/networks';
 import {
     GasSpeed,
     SET_DEFAULT_GAS_SPEED,
-    CHANGE_NETWORK,
 } from './actions';
 
 const initialState: SettingsState = {
@@ -12,7 +10,6 @@ const initialState: SettingsState = {
         speed: 'fast',
         confidence: 99,
     },
-    currentNetworkChainId: MAINNET.chainID,
     loading: false,
     error: null,
 };
@@ -27,11 +24,6 @@ const settingsReducer = (
                 ...state,
                 defaultGasSpeed: action.payload,
             };
-        case CHANGE_NETWORK:
-            return {
-                ...state,
-                currentNetworkChainId: action.payload,
-            };
         default:
             return state;
     }
@@ -40,7 +32,6 @@ const settingsReducer = (
 export type SettingsState = {
     defaultGasSpeed: GasSpeed;
     loading: boolean;
-    currentNetworkChainId: number;
     error: Error | null;
 }
