+++ crosscash-extension/src/popup/store/wallet/reducer.ts
@@ -1,17 +1,23 @@
 import { PayloadAction } from '@reduxjs/toolkit';
 
+import { MAINNET } from '../../../lib/constants/networks';
 import { EthereumWallet } from '../../model/wallet';
-import { createWallet } from './actions';
+import {
+    createWallet,
+    CHANGE_NETWORK,
+    WalletPayloadAction,
+} from './actions';
 
 const initialState: WalletState = {
     walletInstance: null,
+    currentNetworkChainId: MAINNET.chainID,
     loading: false,
     error: null,
 };
 
 export const walletReducer = (
     state = initialState,
-    action: PayloadAction<EthereumWallet & Error>,
+    action: PayloadAction<WalletPayloadAction>,
 ): WalletState => {
     switch (action.type) {
         case createWallet.TRIGGER:
@@ -31,6 +37,11 @@ export const walletReducer = (
                 ...state,
                 loading: false,
             };
+        case CHANGE_NETWORK:
+            return {
+                ...state,
+                currentNetworkChainId: action.payload,
+            };
         default:
             return state;
     }
@@ -38,6 +49,7 @@ export const walletReducer = (
 
 export interface WalletState {
     walletInstance: EthereumWallet | null;
+    currentNetworkChainId: number;
     loading: boolean;
     error: Error | null;
 }
